fix(paid-access): critical generation paywall bypass + review fixes, polish, missing migration
Follow-up to the phase-1 cutover, from a bug-hunt review pass plus readability work.

Bug fixes (a review found these were wrong on their own terms):
- CRITICAL: non-buyers could generate for free with paywalled models. Gated versions
  are availability='Public' now, so the generation path's resource-data cache set
  hasAccess=true and skipped all paid-access enforcement. applyPaidAccessGating now
  demotes/re-decides access via getPaidAccess (owner/mod, purchased, or free/trial).
- donateToGoal no longer refunds a committed donation when the post-commit completion
  writes throw (guard checkDonationGoalComplete, matching earlyAccessPurchase).
- trialLimit/freePreviewGenerations require .int() (a fractional value crashed the
  mini/[id] raw-SQL CAST with a 500).
- assertPaidAccessInput rejects a generation-only paid tier with no price and no
  download tier (earlyAccessPurchase would otherwise charge `undefined` Buzz).
- Clearing a gate resets availability EarlyAccess -> Public, so a version migrated at
  cutover isn't left permanently locked by hasEntityAccess.

Missing migration (schema/code referenced columns the create_paid_access migration
omitted): 20260727140000_paid_access_create_omissions adds PaidAccess.timeframeDays
(+ backfill) and the [entityType, endsAt] index used by the expiry job/notification.

Polish (complexity-review items): unify the triplicated EA caps into
assertUserEarlyAccessLimits; hasEntityAccess skips getPaidAccess for moderators +
grantAll dedup; isTimedGateActive in getDonationGoals; donationGoalLocked in the form;
buyer-path + accessor-contract comments.

generation.service: collapse the demote/lookup/re-grant dance into one
applyPaidAccessGating; the purchase lookup returns a Set (O(1), was an O(n^2) find);
the access rule is shared free functions in @civitai/buzz (grantsGeneration /
generationOpenToNonBuyers / generationTrialLimit) over plain data — no class on the
wire, serialization-safe.

+++ src/server/schema/model-version.schema.ts
@@ -368,9 +368,11 @@ const downloadGrantSchema = z.object({ price: z.number().min(100).max(MAX_DONATI
 const generationGrantSchema = z.union([
   z.object({ free: z.literal(true) }),
   // `price` optional → falls back to the download price; `trialLimit` = free test generations.
+  // trialLimit must be an integer: mini/[id].ts CASTs it to int in raw SQL, which hard-errors (500)
+  // on a fractional value rather than rounding.
   z.object({
     price: z.number().min(50).optional(),
-    trialLimit: z.number().min(0).max(1000).optional(),
+    trialLimit: z.number().int().min(0).max(1000).optional(),
   }),
 ]);
 export const modelVersionTermsSchema = z.object({
